move icon components
