+++ frontend-two/components/icons/FrequentIcon.tsx
