+++ frontend-two/components/icons/timingPointIcons.ts
